:sparkles: Feat: 뷰모델에 프로필 페이지에서 사용할 임시 유저 객체를 선언

+++ app/src/main/java/com/example/snack4diet/api/ApiModels.kt
@@ -2,8 +2,8 @@ package com.example.snack4diet.api
 
 data class UserInfo(
     val nickname: String,
-    val height: Float,
-    val weight: Float,
+    val height: Double,
+    val weight: Double,
     val sex: Boolean,
     val age: Int
 )
